Move internal Theater classes to support package
