+++ org-code-javabuilder/theater/src/main/java/org/code/theater/support/TheaterMessage.java
@@ -1,11 +1,11 @@
-package org.code.theater;
+package org.code.theater.support;
 
 import java.util.HashMap;
 import org.code.protocol.ClientMessage;
 import org.code.protocol.ClientMessageType;
 
 public class TheaterMessage extends ClientMessage {
-  TheaterMessage(TheaterSignalKey key, HashMap<String, String> detail) {
+  public TheaterMessage(TheaterSignalKey key, HashMap<String, String> detail) {
     super(ClientMessageType.THEATER, key.toString(), detail);
   }
 }
