+++ org-code-javabuilder/theater/src/test/java/org/code/theater/support/TheaterProgressPublisherTest.java
@@ -1,4 +1,4 @@
-package org.code.theater;
+package org.code.theater.support;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
